ai-project-assistant/requirements.txt
streamlit>=1.18.0
pandas>=2.0.0
numpy>=1.22.0
plotly>=5.0.0
langchain>=0.0.300
openai>=0.27.0
python-dotenv>=1.0.0
pdfkit>=1.0.0   # optional (requires wkhtmltopdf on system) 
